hw4/requirements.txt
####### requirements.txt #######

###### Requirements without Version Specifiers ######


###### Requirements with Version Specifiers ######

numpy >= 1.12.1
sklearn >= 0.0
Pillow >= 3.1.2

###### Refer to other requirements files ######

###### A particular file ######

###### Additional Requirements without Version Specifiers ######

